requirements.txt
discord.py
python-dotenv
huggingface_hub
feedparser
aiohttp